+++ STM32F407VET6/recorder/Core/User/W25QXX/w25qxx.h
@@ -61,9 +61,9 @@ void W25QXX_4ByteAddr_Enable(void);     //使能4字节地址模式
 void W25QXX_Write_SR(uint8_t regno,uint8_t sr);   //写状态寄存器
 void W25QXX_Write_Enable(void);  		//写使能 
 void W25QXX_Write_Disable(void);		//写保护
-void W25QXX_Write_NoCheck(uint8_t* pBuffer,uint32_t WriteAddr,uint16_t NumByteToWrite);
-void W25QXX_Read(uint8_t* pBuffer,uint32_t ReadAddr,uint16_t NumByteToRead);   //读取flash
-void W25QXX_Write(uint8_t* pBuffer,uint32_t WriteAddr,uint16_t NumByteToWrite);//写入flash
+void W25QXX_Write_NoCheck(uint8_t* pBuffer,uint32_t WriteAddr,uint32_t NumByteToWrite);
+void W25QXX_Read(uint8_t* pBuffer,uint32_t ReadAddr,uint32_t NumByteToRead);   //读取flash
+void W25QXX_Write(uint8_t* pBuffer,uint32_t WriteAddr,uint32_t NumByteToWrite);//写入flash
 void W25QXX_Erase_Chip(void);    	  	//整片擦除
 void W25QXX_Erase_Sector(uint32_t Dst_Addr);	//扇区擦除
 void W25QXX_Wait_Busy(void);           	//等待空闲
